feat(06-02): migrate cost_basis_aud/usd from float to Decimal (HARD-06)
- models.py: OpenLot and DisposedLot cost_basis fields now Decimal
- ledger.py: _FLOAT_TOLERANCE=Decimal('0.001'), Decimal proportion arithmetic
- engine.py: OpenLot construction uses Decimal(str(cost_aud/usd))
- engine.py: _build_disposed_lot_with_cgt converts Decimal to float for gain
- audit.py: float(lot.cost_basis_aud) at CgtEventRow construction site
- test_tax_ledger.py: all fixtures use Decimal; 4 new type-assertion tests
- test_tax_cgt.py: _make_lot and fixtures updated to use Decimal cost_basis_aud
- test_tax_engine.py: float() conversion at cost_basis_aud assertions
- mypy --strict: clean across all modified files

+++ src/market_data/backtest/tax/models.py
@@ -10,6 +10,7 @@ from __future__ import annotations
 import io
 from dataclasses import dataclass
 from datetime import date
+from decimal import Decimal
 
 from rich.console import Console, ConsoleOptions, RenderResult
 from rich.table import Table
@@ -27,8 +28,8 @@ class OpenLot:
     ticker: str
     acquired_date: date
     quantity: float
-    cost_basis_aud: float
-    cost_basis_usd: float | None
+    cost_basis_aud: Decimal
+    cost_basis_usd: Decimal | None
 
 
 @dataclass(frozen=True)
@@ -43,8 +44,8 @@ class DisposedLot:
     acquired_date: date
     disposed_date: date
     quantity: float
-    cost_basis_usd: float | None
-    cost_basis_aud: float
+    cost_basis_usd: Decimal | None
+    cost_basis_aud: Decimal
     proceeds_usd: float | None
     proceeds_aud: float
     gain_aud: float
